+++ Formula/millennium-helpers.rb
@@ -1,7 +1,7 @@
 class MillenniumHelpers < Formula
   desc "Cross-platform utilities and MCP server for Millennium Steam Client hook"
   homepage "https://github.com/bolens/millenium-helpers"
-  url "https://github.com/bolens/millenium-helpers/releases/download/v2.6.1/millennium-helpers-linux.tar.gz"
+  url "https://github.com/bolens/millenium-helpers/releases/download/v2.6.2/millennium-helpers-linux.tar.gz"
   sha256 "81c20635326ea945f23729f6c47021a0ba2bd4dfccf1b367bcd48dfdee363827"
   license "MIT"
   head "https://github.com/bolens/millenium-helpers.git", branch: "main"